Moved all send/receive topic filter tests to topicfilter file.
